// cosmowcl/src/app/shared/enums/nome-funzionalita.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
export enum NomeFunzionalita {

  CONSULTAZIONE_DOCUMENTI = 'CONSULTAZIONE-DOCUMENTI',
  APPROVAZIONE = 'APPROVAZIONE',
  PIANIFICAZIONE_ATTIVITA = 'SELEZIONE-APPROVATORI',
  GESTIONE_DOCUMENTI = 'GESTIONE-DOCUMENTI',
  COMMENTI = 'COMMENTI',
  SIMPLE_FORM = 'SIMPLE-FORM',
  COLLABORAZIONE = 'COLLABORAZIONE',
  FIRMA_DOCUMENTI = 'FIRMA-DOCUMENTI',
  ATTIVA_SISTEMA_ESTERNO = 'ATTIVA-SISTEMA-ESTERNO',
  ASSOCIAZIONE_PRATICHE = 'ASSOCIAZIONE-PRATICHE',
  GENERAZIONE_REPORT = 'GENERAZIONE-REPORT',
  CUSTOM_FORM = 'CUSTOM-FORM',
  CREAZIONE_PRATICA = 'CREAZIONE-PRATICA',
  SCELTA = 'SCELTA',
  ASSEGNAZIONE_TAGS = 'ASSEGNAZIONE-TAGS'
}